requirements.txt
praw>=7.7.1
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
pytz>=2023.3
tqdm>=4.65.0
pyyaml>=6.0.1
matplotlib>=3.8.0
